HTTP GET requests using Node.js
HTTP-WebServer (Node.js)

A web-server created with the Nodejs HTTP Module that responds to two GET requests, one to list all information on jobs in a JSON array, while the second one is to find information about the quantity required for a specific job and part from a specific JSON object. Information stored about the jobs is stored in an array of objects, such that each object has three attributes: jobName, partId, qty.